packages/data-pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
httpx>=0.25.0
supabase>=2.0.0
tenacity>=8.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
